+++ src/ce/ti/debug.h
@@ -1,0 +1,31 @@
+/**
+ * @file
+ * @authors
+ * Matt "MateoConLechuga" Waltz\n
+ * Jacob "jacobly" Young
+ * @brief This file is pretty much pointless.
+ */
+
+#ifndef _TI_DEBUG_H
+#define _TI_DEBUG_H
+
+#include <stdbool.h>
+#include <stdint.h>
+#include <stddef.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/**
+ * Basically a reimplemented form of printf that prints to some debugging device
+ *
+ * @param string String to send to debug device
+ */
+void boot_DebugPrintf(const char *string);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
